requirements.txt
librosa>=0.10.0
numpy>=1.24.0
tensorflow>=2.13.0
tensorflow-hub>=0.15.0
scikit-learn>=1.3.0
pyyaml>=6.0
modal>=0.64.0
sounddevice>=0.4.6
requests>=2.31.0
ably>=2.0.0
boto3>=1.34.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
